notification module test
