+++ controllers/notification/config/crd
@@ -1,0 +1,1 @@
+/root/sealos/controllers/common/notification/config/crd
